Update the "more verbose" tutorial example
There are two changes here, one of which may be a bit contentious. I
really enjoyed working through this, but I was thrown by the final
comments concerning 167. What's so different about 167 compared to 168?
I found the program generator was quite happy to come up with a program
for both.

I thought it might be interesting therefore if the 167 example was
something that the program generator was not able to solve. This, I
think, makes it more interesting and also emphasises that the software
is doing what the user would expect (i.e. not coming up with a solution
that doesn't exist).

This change therefore constrains the values to being even, so that the
167 result can't be obtained.

The second, smaller change, is to use 'Number# instead of 'Real'. I was
uncomfortable with the latter simply because, even in the original
version, they are in fact all integers.

I understand if you feel these changes are making things overly
complicated. I just thought they'd make an interesting example.