chore: introduce monorepo structure
Switching to a monorepo has long been on my wishlist for Herb,
and now that Pkg added support for workspaces and added support
for depending on packages within a monorepo, the necessary tooling
is in place for us to make the switch.

The reason for this change is that it will allow us to spend a lot less
time on updates that involve multiple packages, updates that involve
documentation, and updates whose changes might effect multiple packages.

For changes that involve multiple packages, especially dependency compat
bumps, we currently have to make PRs in each of the repositories, and
then make a release in each repository, starting with the deepest
affected Herb* package in the dependency tree, and then work our way up.
For each release, we wait for the registry action to pass and then we
register the next package. It would be MUCH cleaner if we could just
make a single PR updating all of the relevant compat bounds, check that
tests pass in all packages (and that no benchmarks are running slower,
once we include benchmarks for at least some of the repositories!!), and
then release new versions for all of the relevant packages
(JuliaRegistrator can register a subdirectory, allowing this workflow).

For changes that involve documentation, the current workflow is
extremely convoluted, which I think contributes to a lack of updates and
lack of coverage. Right now, if I change some docstrings for a
particular Herb* package, the only way I can see the updated
documentation is by cloning the Herb.jl repo, `dev`ing the repo where
I've made the changes to documentation, and then building the
documentation in the Herb.jl repo. Ideally, I'd like any PR that makes
changes that affect documentation to have a comment from the CI with a
link to the documentation preview. This would be MUCH easier if we have
a single CI setup for all of the repos. As a side note, having a single
`.github/workflows` setup would make any changes to CI much easier. We
currently have to make a PR in every repo with changes.

The last point is that changes in, for example, HerbCore, might affect
any of the other Herb* packages. There's no automated check for this at
the moment, so if we introduce a change that introduces some
(performance) regression in another package, we never find out about it
until after we've merged and released. Ideally, we should set up the CI
such that changes in, for example, HerbCore, trigger the test suites of
packages that depend on it. Once we set up benchmarks, we should trigger
the benchmarks too.

All of these changes allow us to keep each of the Herb* packages as its
own package, meaning users can install them on their own, just like now.
For the end user, this change should go completely unnoticed--except for
the those visiting the GitHub pages, as all Herb* would all be under
(this) one repository.

For the moment, I've included these as subrepos, so that we can stay up
to date with the changes while this PR is in draft form, but ideally,
we'd want to have them as normal subdirectories. We should also try to
keep their history if possible, so that we don't lose git blame
information upon merging.